rename ObjectIntMappingFactory
